+++ __mocks__/@react-native-clipboard/clipboard.js
@@ -1,0 +1,14 @@
+let clipboardContents = '';
+
+const Clipboard = {
+  getString: jest.fn(async () => clipboardContents),
+  setString: jest.fn(content => {
+    clipboardContents = content;
+  }),
+  __reset: () => {
+    clipboardContents = '';
+  },
+};
+
+module.exports = Clipboard;
+module.exports.default = Clipboard;